+++ packages/api/jest.int.config.js
@@ -3,4 +3,5 @@ module.exports = {
   modulePathIgnorePatterns: ['<rootDir>/dist'],
   testMatch: ['<rootDir>/test/integration/**/*.test.ts'],
   collectCoverageFrom: ['src/**/*.{ts,tsx,js,jsx}'],
+  fakeTimers: { enableGlobally: true },
 }
