+++ packages/graphql-schema/src/typeDefs/advertisement.graphql
