+++ packages/graphql-schema/src/typeDefs/aggregateOffer.graphql
