+++ packages/graphql-schema/src/typeDefs/aggregateRating.graphql
