+++ packages/graphql-schema/src/typeDefs/facet.graphql
