+++ packages/graphql-schema/src/typeDefs/image.graphql
