+++ packages/graphql-schema/src/typeDefs/newsletter.graphql
