+++ packages/graphql-schema/src/typeDefs/offer.graphql
