+++ packages/graphql-schema/src/typeDefs/product.graphql
