+++ packages/graphql-schema/src/typeDefs/productGroup.graphql
