+++ packages/graphql-schema/src/typeDefs/propertyValue.graphql
