+++ packages/graphql-schema/src/typeDefs/skuVariants.graphql
