+++ packages/graphql-schema/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "resolveJsonModule": true,
+    "moduleResolution": "Bundler"
+  },
+  "include": ["src/**/*"]
+}
