+++ packages/graphql-schema/vite.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vite'
+import dts from 'vite-plugin-dts'
+
+export default defineConfig({
+  root: process.env.PWD ?? process.cwd(),
+  plugins: [dts()],
+  build: {
+    lib: {
+      entry: './src/index.ts',
+      formats: ['cjs', 'es'],
+      fileName: '[format]/index',
+    },
+    rollupOptions: {
+      external: ['graphql'],
+    },
+  },
+})
