+++ pnpm-workspace.yaml
@@ -1,32 +1,35 @@
 packages:
-  - "packages/*"
-  - "apps/*"
+  - packages/*
+  - apps/*
 
 catalog:
-  "@antfu/ni": "^0.21.12"
-  "@graphql-codegen/cli": "5.0.2"
-  "@graphql-codegen/typescript": "4.0.7"
-  "@graphql-tools/merge": "^9.1.1"
-  "@graphql-tools/schema": "^10.0.3"
-  "@graphql-tools/utils": "^10.2.0"
-  "@jest/globals": "29.7.0"
-  "@testing-library/react": "^14.3.0"
-  "@types/jest": "29.1.0"
-  "@types/node": "^20"
-  "@types/react-dom": "^19"
-  "@types/react": "^19"
-  "@types/tabbable": "^3.1.1"
-  "cookie": "^0.7.0"
-  "fast-deep-equal": "^3.1.3"
-  "graphql": "15.6.0"
-  "jest-environment-jsdom": "^29.7.0"
-  "jest": "^29.7.0"
-  "next": "^15.5.2"
-  "react-dom": "^19"
-  "react-intersection-observer": "^8.32.5"
-  "react-swipeable": "^7.0.0"
-  "react": "^19"
-  "tabbable": "^5.2.1"
-  "ts-jest": "29.1.1"
-  "tslib": "^2.3.1"
-  "typescript": "5.3.2"
+  '@antfu/ni': ^0.21.12
+  '@graphql-codegen/cli': 5.0.2
+  '@graphql-codegen/typescript': 4.0.7
+  '@graphql-tools/merge': 9.1.1
+  '@graphql-tools/schema': ^10.0.3
+  '@graphql-tools/utils': ^10.2.0
+  '@jest/globals': 29.7.0
+  '@testing-library/react': ^14.3.0
+  '@types/jest': 29.1.0
+  '@types/node': ^20
+  '@types/react': ^19
+  '@types/react-dom': ^19
+  '@types/tabbable': ^3.1.1
+  cookie: ^0.7.0
+  fast-deep-equal: ^3.1.3
+  graphql: 15.6.0
+  jest: ^29.7.0
+  jest-environment-jsdom: ^29.7.0
+  next: ^15.5.2
+  react: ^19
+  react-dom: ^19
+  react-intersection-observer: ^8.32.5
+  react-swipeable: ^7.0.0
+  tabbable: ^5.2.1
+  ts-jest: 29.1.1
+  tslib: ^2.3.1
+  typescript: 5.3.2
+
+overrides:
+  graphql: 15.6.0
